+++ sources/packages/backend/apps/api/src/route-controllers/education-program/education-program.controller.service.ts
@@ -214,6 +214,7 @@ export class EducationProgramControllerService {
       institutionProgramCode: program.institutionProgramCode,
       minHoursWeek: program.minHoursWeek,
       isAviationProgram: program.isAviationProgram,
+      credentialTypesAvi: program.credentialTypesAviation,
       minHoursWeekAvi: program.minHoursWeekAvi,
       entranceRequirements: {
         hasMinimumAge: program.hasMinimumAge,
