+++ sources/packages/backend/apps/api/src/services/education-program/education-program.service.ts
@@ -225,6 +225,7 @@ export class EducationProgramService extends RecordDataModelService<EducationPro
       program.institutionProgramCode = educationProgram.institutionProgramCode;
       program.minHoursWeek = educationProgram.minHoursWeek;
       program.isAviationProgram = educationProgram.isAviationProgram;
+      program.credentialTypesAviation = educationProgram.credentialTypesAvi;
       program.minHoursWeekAvi = educationProgram.minHoursWeekAvi;
       program.minHighSchool =
         educationProgram.entranceRequirements.minHighSchool;
@@ -538,6 +539,7 @@ export class EducationProgramService extends RecordDataModelService<EducationPro
         "programs.institutionProgramCode",
         "programs.minHoursWeek",
         "programs.isAviationProgram",
+        "programs.credentialTypesAviation",
         "programs.minHoursWeekAvi",
         "programs.hasMinimumAge",
         "programs.minHighSchool",
