+++ sources/packages/backend/apps/db-migrations/src/sql/EducationPrograms/Rollback-add-credential-types-aviation-column.sql
@@ -1,0 +1,5 @@
+ALTER TABLE
+  sims.education_programs DROP COLUMN credential_types_aviation;
+
+ALTER TABLE
+  sims.education_programs_history DROP COLUMN credential_types_aviation;
