+++ sources/packages/backend/apps/db-migrations/src/sql/EducationPrograms/Rollback-update-is-aviation-program-to-no.sql
@@ -1,0 +1,9 @@
+UPDATE
+  sims.education_programs
+SET
+  is_aviation_program = NULL
+WHERE
+  course_load_calculation = 'credit' || (
+    course_load_calculation = 'hours'
+    AND min_hours_week = 'yes'
+  );
