+++ sources/packages/backend/apps/db-migrations/src/sql/EducationPrograms/Update-is-aviation-program-to-no.sql
@@ -1,0 +1,6 @@
+UPDATE
+  education_programs
+SET
+  is_aviation_program = 'no'
+WHERE
+  is_aviation_program IS NULL;
