+++ sources/packages/backend/libs/sims-db/src/entities/education-program.model.ts
@@ -219,6 +219,18 @@ export class EducationProgram extends RecordDataModel {
   isAviationProgram?: string;
 
   /**
+   * A jsonb column to store various credential types related to aviation programs.
+   */
+  @Column({
+    name: "credential_types_aviation",
+    type: "jsonb",
+    nullable: true,
+    comment:
+      "A jsonb column to store various credential types related to aviation programs.",
+  })
+  credentialTypesAviation?: AviationProgramCredentialTypes;
+
+  /**
    * Minimum hours check for a aviation program.
    */
   @Column({
@@ -430,4 +442,15 @@ export class EducationProgram extends RecordDataModel {
     nullable: true,
   })
   isActiveUpdatedOn?: Date;
+}
+
+/**
+ * Aviation program credential types.
+ */
+export interface AviationProgramCredentialTypes {
+  commercialPilotTraining: boolean;
+  instructorsRating: boolean;
+  endorsements: boolean;
+  privatePilotTraining: boolean;
+  noneOfTheAbove: boolean;
 }
